+++ src/locales/en/contact-form.js
@@ -4,8 +4,8 @@
 export default {
   contactFormHoneypot: 'Don’t fill this out if you’re a human',
   contactFormName: 'Name',
-  contactFormEmail: 'Email',
   contactFormEither: 'At least either',
+  contactFormEmail: 'Email',
   contactFormOr: 'or',
   contactFormTel: 'Phone Number',
   contactFormMessage: 'Your message for us',